+++ docker-compose.production.yml
@@ -27,7 +27,7 @@ services:
       - capstone-backend
 
   capstone:
-    image: dihson103/capstone:v5.17
+    image: dihson103/capstone:v5.18
     container_name: capstone
     environment:
       - ASPNETCORE_ENVIRONMENT=Development
