+++ src/Contract/Services/ShipOrder/AcceptShipOrder/AcceptShipOrderCommand.cs
@@ -1,0 +1,5 @@
+﻿using Contract.Abstractions.Messages;
+
+namespace Contract.Services.ShipOrder.AcceptShipOrder;
+
+public record AcceptShipOrderCommand(Guid shipOrderId, string updatedBy) : ICommand;
