+++ src/Contract/Services/ShipOrder/GetShipOrderByOrderId/ShipOrderResponse.cs
@@ -8,6 +8,7 @@ public record ShipOrderResponse(
     string ShipperId, 
     string ShipperName,
     DateTime ShipDate,
+    bool IsAccepted,
     Status Status,
     string StatusDescription,
     DeliveryMethod DeliveryMethod,
